// backend/src/common/types/pagination.type.ts
export type QueryType = {
    page?: string;
    limit?: string;
    search?: string;
    sortedBy?: string;
    orderBy?: "asc" | "desc";
};

export type ListingQueryType = {
    minPrice?: string;
    maxPrice?: string;
    currency?: string;
    userId?: string;
    isProfanity?: string | boolean;
    isActive?: string | boolean;
} & QueryType;

export type UserQueryType = {
    userId?: string;
    isActive?: string;
} & QueryType;

export type PaginatedResponseType<T> = {
    docs: T[];
    totalDocs: number;
    offset: number;
    limit: number;
    totalPages: number;
    page?: number;
    pagingCounter?: number;
    hasPrevPage: boolean;
    hasNextPage: boolean;
    prevPage?: number | null;
    nextPage?: number | null;
};

export type PaginateOptionsType = {
    page?: number;
    limit?: number;
    sort?: { [key: string]: 1 | -1 };
    select?: string | string[];
    populate?: string | any;
    lean?: boolean;
};

export type PaginateFilterType = {
    $or?: Array<{ [key: string]: any }>;
    [key: string]: any;
};